Deletes the older one and add recursion to the newer.
