+++ src/main/java/algorithms/curated170/medium/candycrush/CandyCrush.java
@@ -1,6 +1,4 @@
 package algorithms.curated170.medium.candycrush;
- 
-import java.util.Arrays;
 
 public class CandyCrush {
 
@@ -12,132 +10,79 @@ public class CandyCrush {
             {1, 2},
             {6, 2},
             {6, 2},
-            {1, 2},
-           
-        };
-        
-        candyCrush(board);
-   
+            {1, 2},};
+
+        CandyCrush candyCrush = new CandyCrush();
+        candyCrush.candyCrush(board);
+
     }
-  
-     static boolean  signal = false; 
-      
-      public static int[][] candyCrush(int[][] board) {
-        int[][] boardCopy = marking(board);
-        
-        System.out.println("DROP:");
-        board =  drop(boardCopy);
-         
-         marking(board);
-         
-        return signal ? candyCrush(board) : board;
 
-        
-         
+    public int[][] candyCrush(int[][] board) {
+        int[][] target = new int[board.length][board[0].length];
+
+        if (crush(board, target)) {
+            int[][] temp = board;
+            board = target;
+            target = temp;
+            return candyCrush(board);
+        } else {
+            return board;
+        }
+
     }
-     
-    public static int[][] marking(int[][] board){
-        signal = false;
-        int[][] boardCopy = new int[board.length][board[0].length];
 
-        for (int i = 0; i < board.length; i++) {
-            int low = 0;
+    private boolean crush(int[][] source, int[][] target) {
+        copy(source, target);
 
-            for (int high = 1; high < board[i].length + 1; high++) {
+        boolean crush = false;
+        for (int i = 0; i < source.length; i++) {
+            int[] sourceRow = source[i];
+            int[] targetRow = target[i];
 
-                while (high < board[i].length && board[i][low] != board[i][high] 
-                        || low < high && high == board[i].length) {
-
-                    if (board[i][low]!=0 && high - low  >= 3) {
-                        
-                        signal = true;
-                        while (low <high) {
-                            boardCopy[i][low] = 0;
-                            low++;
-
-                        }
-                    } else {
-                        while (low <high) {
-
-                            boardCopy[i][low] = board[i][low];
-                            low++;
-
-                        }
-
+            for (int j = 0; j < sourceRow.length; j++) {
+                if (j < sourceRow.length - 2) {
+                    if (sourceRow[j] != 0 && sourceRow[j] == sourceRow[j + 1] && sourceRow[j] == sourceRow[j + 2]) {
+                        crush = true;
+                        targetRow[j] = 0;
+                        targetRow[j + 1] = 0;
+                        targetRow[j + 2] = 0;
                     }
-
                 }
-
-            }
-
-        }
-        
-        
-        for (int j = 0; j < board[0].length; j++) {
-            int low = 0;
-
-            for (int high = 1; high < board.length + 1; high++) {
-
-                while (high < board.length && board[low][j] != board[high][j] 
-                        || low < high && high == board.length) {
-
-                    if (board[low][j]!=0 && high - low  >= 3) {
-                        signal = true;
-                        while (low <high) {
-                            boardCopy[low][j] = 0;
-                            low++;
-
-                        }
-                    } else {
-                        while (low <high) {
-                                
-                 
-                            low++;
-
-                        }
-
+                if (i < source.length - 2) {
+                    if (sourceRow[j] != 0 && sourceRow[j] == source[i + 1][j] && sourceRow[j] == source[i + 2][j]) {
+                        crush = true;
+                        targetRow[j] = 0;
+                        target[i + 1][j] = 0;
+                        target[i + 2][j] = 0;
                     }
-
-                }
-
-            }
-
-        }
-       return boardCopy;
-    }  
-  
-    
-    
-    public static int[][] drop( int[][] array) { 
-
-       
-        int[][] arrayClone = new int[array.length][array[0].length];
-
-  
-
-        for (int j = 0; j < array[0].length; j++) {
-            
-            int counter = array.length - 1;
-            
-            for (int i = array.length - 1; i >= 0; i--) {
-
-                if (array[i][j] != 0) {
-                    arrayClone[counter--][j] = array[i][j];
-
                 }
             }
         }
-        printArray(arrayClone);
-     
-        return arrayClone; 
+
+        if (crush) {
+            drop(target);
+        }
+        return crush;
     }
 
+    private void drop(int[][] target) {
+        for (int col = 0; col < target[0].length; col++) {
+            int targetIndex = target.length - 1;
+            for (int row = target.length - 1; row >= 0; row--) {
+                if (target[row][col] != 0) {
+                    if (row != targetIndex) {
+                        target[targetIndex][col] = target[row][col];
+                        target[row][col] = 0;
+                    }
+                    targetIndex--;
+                }
+            }
+        }
+    }
 
-    
-    static void printArray(int[][] boardToPrint){
-        for (int i = 0; i < boardToPrint.length; i++) {
-            System.out.println( Arrays.toString(boardToPrint[i]));
-          
+    private void copy(int[][] source, int[][] target) {
+        for (int i = 0; i < source.length; i++) {
+            System.arraycopy(source[i], 0, target[i], 0, source[i].length);
         }
     }
 }
